fix: import dart:math and cast distKm in api_service.dart

+++ mobile/lib/services/api_service.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:math';
 import 'package:flutter/foundation.dart' show kIsWeb, kReleaseMode;
 import 'package:http/http.dart' as http;
 import 'package:supabase_flutter/supabase_flutter.dart';
@@ -458,7 +459,7 @@ class ApiService {
       const p = 0.017453292519943295;
       final a = 0.5 - cos((lat2 - lat1) * p) / 2 +
           cos(lat1 * p) * cos(lat2 * p) * (1 - cos((lon2 - lon1) * p)) / 2;
-      return 12742 * asin(sqrt(a));
+      return (12742 * asin(sqrt(a))).toDouble();
     }
 
     final seenKeys = <String>{};
